refactor: error utils
