--- packages/apps/reputation-oracle/server/src/common/enums/minio.ts
@@ -1,3 +1,0 @@
-export enum MinioErrorCodes {
-  NotFound = 'NotFound',
-}
