--- packages/apps/reputation-oracle/server/src/common/utils/database.ts
@@ -1,9 +1,0 @@
-import { PostgresErrorCodes } from '../enums/database';
-import { DatabaseError } from '../errors/database';
-
-export function isDuplicatedError(error: any): boolean {
-  return (
-    error instanceof DatabaseError &&
-    error.message.includes(PostgresErrorCodes.Duplicated)
-  );
-}
