+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-incoming.service.ts
@@ -5,7 +5,7 @@ import { Web3ConfigService } from '../../common/config/web3-config.service';
 import { BACKOFF_INTERVAL_SECONDS } from '../../common/constants';
 import { EventType, WebhookIncomingStatus } from '../../common/enums';
 import { calculateExponentialBackoffMs } from '../../utils/backoff';
-import { isDuplicatedError } from '../../common/utils/database';
+import { isDuplicatedError } from '../../common/errors/database';
 import { EscrowCompletionService } from '../escrow-completion/escrow-completion.service';
 import { WebhookIncomingEntity } from './webhook-incoming.entity';
 import { WebhookIncomingRepository } from './webhook-incoming.repository';
